uuid: check that uuid file contains valid uuid (#3426)
Currently, if the uuid file gets corrupted, BareSIP gets stuck in a loop
where it keeps trying to use an invalid uuid instead of generating a new
one. Valid UUIDs are needed for example when SIP outbound is used with
an account, so in this scenario, registrations to a SIP server are
broken and stay broken until someone manually deletes the uuid file.

This is fixed now by checking that the uuid file contains a valid uuid.

Upstream-Status: Backport [ac26c5b62c9946c2ebd3653fac7235f82c389741]
Signed-off-by: Matt Madison <matt@madison.systems>